2017-04-03 release
Changes to the KSR Signer “ksrsigner” component:

1. ksrsigner/ksrcommon.c:
a. Added support to process the configuration file “kskschedule.json”.
b. Added support to process the configuration options to sign, publish
or revoke the KSK.

2. ksrsigner/ksrcommon.h:
a. Added definition of json key schedule parsing support.

3. ksrsigner/ksrsigner.c:
a. Added support to locate the configuration file kskschedule.json in the same
place as the KSR.xml file.

4. ksrsigner/wksr.c:
a. Added support to locate the configuration file kskschedule.json in the same
place as the KSR.xml file.

Other changes:

5. utils/printlog:
a. Removed top margin.
b. Added header with name and page.
c. Reduced font size to fit better on the page.

6. utils/hsmfd-hash:
a. New bash script to calculate, print and compare hashes for HSMFDs.

7. ChangeLog:
a. Added a change log file.

8. README.md:
a. Update copyright year to 2017